HealthCare-Management-System

+++ src/main/java/com/example/demo/Appointment/AppointmentDto.java
@@ -9,8 +9,8 @@ import java.time.LocalDate;
 
 @Data
 public class AppointmentDto {
-    Integer patientId;
-    Integer doctorId;
+    Long patientId;
+    Long doctorId;
     @Enumerated(EnumType.STRING)
     private AppointmentStatus status;
     private String notes;
